src/requirements.txt
Adafruit-Blinka==8.57.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-dht==4.0.7
adafruit-circuitpython-mlx90614==1.2.20
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.78.0
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
assure==0.0.6
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
beautifulsoup4==4.13.4
binho-host-adapter==0.1.6
bs4==0.0.2
cbor2==5.6.5
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
drivers==0.0.5
executing==2.2.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
filelock==3.18.0
fleep==1.0.1
fsspec==2025.3.2
future==1.0.0
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
libarchive-c==5.2
luma.core==2.4.2
luma.oled==3.14.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mmry==0.0.9
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pexpect==4.9.0
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==10.2.0
pluggy==1.5.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyfingerprint==1.5
pyftdi==0.56.0
Pygments==2.19.1
PyMLX90614==0.0.4
pypdf==5.4.0
pyserial==3.5
pytest==8.3.5
python-prctl==1.8.1
pyusb==1.3.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
RPi.GPIO==0.7.1
RPLCD==1.4.0
safetensors==0.5.3
scipy==1.15.2
simplejpeg==1.8.2
smbus==1.1.post2
smbus2==0.5.0
soupsieve==2.7
spidev==3.7
stack-data==0.6.3
sympy==1.14.0
sysv-ipc==1.1.0
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
v4l2-python3==0.3.5
validators==0.35.0
wcwidth==0.2.13
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
